backend/pyproject.toml
[project]
name = "trading-bots-backend"
version = "0.1.0"
description = "Crypto Trading Agents Mission Control - MVP"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0.0",
    "ccxt>=4.0.0",
    "pandas>=2.0.0",
    "ta>=0.11.0",
    "google-adk>=1.0.0",
    "litellm>=1.0.0",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.0.0",
    "websockets>=12.0",
    "psycopg2-binary>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
